minor bug in mvnnorm
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bayesr@2935 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/bamlss/src/bamlss_functions.c
@@ -2971,7 +2971,7 @@ SEXP log_dmvnorm(SEXP Y, SEXP PAR, SEXP N, SEXP K, SEXP MJ, SEXP SJ, SEXP RJ)
   PROTECT(d = allocVector(REALSXP, n));
   double *dptr = REAL(d);
 
-  double lpi = - 0.5 * k * log(2.0 * 3.14159265358979323846);
+  double lpi = - 0.5 * (float)k * log(2.0 * 3.14159265358979323846);
   double det = 0.0;
   double sum = 0.0;
